+++ sql/emaj_prepare_emaj_web_test.sql
@@ -275,11 +275,11 @@ delete from "myTbl3";
 -----------------------------
 
 select emaj.emaj_set_mark_group('myGroup2','tmp_mark');
-select emaj.emaj_logged_rollback_group('myGroup2','MARK1');
-select emaj.emaj_rollback_group('myGroup2','tmp_mark');
+select * from emaj.emaj_logged_rollback_group('myGroup2','MARK1');
+select * from emaj.emaj_rollback_group('myGroup2','tmp_mark');
 select emaj.emaj_delete_mark_group('myGroup2','tmp_mark');
 
-select emaj.emaj_logged_rollback_group('myGroup2','MARK3');
+select * from emaj.emaj_logged_rollback_group('myGroup2','MARK3');
 
 RESET search_path;